+++ pidx/core/PIDX_agg/PIDX_global_agg/PIDX_global_agg.c
@@ -56,7 +56,7 @@ static FILE* agg_dump_fp;
 //static PIDX_return_code one_sided_data_com(PIDX_global_agg_id agg_id, int mode);
 
 #if PIDX_HAVE_MPI
-static PIDX_return_code aggregate_write_read(PIDX_global_agg_id agg_id, int variable_index, uint64_t hz_start_index, uint64_t hz_count, unsigned char* hz_buffer, int buffer_offset, Agg_buffer agg_buffer, PIDX_block_layout local_block_layout, PIDX_block_layout global_block_layout,  int MODE);
+static PIDX_return_code aggregate_write_read(PIDX_global_agg_id agg_id, int variable_index, unsigned long long hz_start_index, unsigned long long hz_count, unsigned char* hz_buffer, int buffer_offset, Agg_buffer agg_buffer, PIDX_block_layout local_block_layout, PIDX_block_layout global_block_layout,  int MODE);
 
 static PIDX_return_code create_open_log_file (PIDX_global_agg_id agg_id);
 
@@ -104,7 +104,7 @@ static PIDX_return_code create_window(PIDX_global_agg_id agg_id, Agg_buffer agg_
 static PIDX_return_code one_sided_data_com(PIDX_global_agg_id agg_id, Agg_buffer agg_buffer, PIDX_block_layout block_layout, int mode)
 {
   int i, p, v, ret = 0, e1 = 0;
-  int64_t index = 0, count = 0;
+  unsigned long long index = 0, count = 0;
   int rank = 0;
   int send_index = 0;
   int hz_index = 0;
@@ -377,7 +377,7 @@ static PIDX_return_code one_sided_file_zero(PIDX_global_agg_id agg_id, PIDX_bloc
   {
     PIDX_variable var = agg_id->idx->variable[v];
     int values_per_sample = var->values_per_sample;
-    int64_t total_chunk_size = agg_id->idx->chunk_size[0] * agg_id->idx->chunk_size[1] * agg_id->idx->chunk_size[2] * agg_id->idx->chunk_size[3] * agg_id->idx->chunk_size[4];
+    unsigned long long total_chunk_size = agg_id->idx->chunk_size[0] * agg_id->idx->chunk_size[1] * agg_id->idx->chunk_size[2] * agg_id->idx->chunk_size[3] * agg_id->idx->chunk_size[4];
     int bytes_per_datatype = ((var->bits_per_value / 8) * total_chunk_size) / (agg_id->idx->compression_factor);
     int target_rank = agg_id->rank_holder[block_layout->inverse_existing_file_index[0]][v - agg_id->first_index][0];
     assert (block_layout->inverse_existing_file_index[0] == 0);
@@ -420,17 +420,17 @@ static PIDX_return_code one_sided_file_zero(PIDX_global_agg_id agg_id, PIDX_bloc
 
 
 #if PIDX_HAVE_MPI
-static PIDX_return_code aggregate_write_read(PIDX_global_agg_id agg_id, int variable_index, uint64_t hz_start_index, uint64_t hz_count, unsigned char* hz_buffer, int buffer_offset, Agg_buffer agg_buffer, PIDX_block_layout local_block_layout, PIDX_block_layout global_block_layout, int MODE)
+static PIDX_return_code aggregate_write_read(PIDX_global_agg_id agg_id, int variable_index, unsigned long long hz_start_index, unsigned long long hz_count, unsigned char* hz_buffer, int buffer_offset, Agg_buffer agg_buffer, PIDX_block_layout local_block_layout, PIDX_block_layout global_block_layout, int MODE)
 {
   int rank = 0, itr;// nrank = 0;
   int bytes_per_datatype;
   int file_no = 0, block_no = 0, negative_block_offset = 0, sample_index = 0, values_per_sample;
   int target_rank = 0;
-  int64_t start_agg_index = 0, end_agg_index = 0, target_disp = 0, target_count = 0, hz_start = 0, samples_in_file = 0;
-  int64_t samples_per_file = (int64_t) agg_id->idx_d->samples_per_block * agg_id->idx->blocks_per_file;
+  unsigned long long start_agg_index = 0, end_agg_index = 0, target_disp = 0, target_count = 0, hz_start = 0, samples_in_file = 0;
+  unsigned long long samples_per_file = (unsigned long long) agg_id->idx_d->samples_per_block * agg_id->idx->blocks_per_file;
   //MPI_Aint target_disp_address;
 
-  int64_t total_chunk_size = agg_id->idx->chunk_size[0] * agg_id->idx->chunk_size[1] * agg_id->idx->chunk_size[2] * agg_id->idx->chunk_size[3] * agg_id->idx->chunk_size[4];
+  unsigned long long total_chunk_size = agg_id->idx->chunk_size[0] * agg_id->idx->chunk_size[1] * agg_id->idx->chunk_size[2] * agg_id->idx->chunk_size[3] * agg_id->idx->chunk_size[4];
 
 #if PIDX_HAVE_MPI
   int ret;
@@ -488,8 +488,8 @@ static PIDX_return_code aggregate_write_read(PIDX_global_agg_id agg_id, int vari
   hz_buffer = hz_buffer + buffer_offset * bytes_per_datatype * values_per_sample;
 #endif
 
-  start_agg_index = target_disp / (int64_t) (samples_in_file / agg_buffer->aggregation_factor);
-  end_agg_index = ((target_disp + target_count - 1) / (int64_t) (samples_in_file / agg_buffer->aggregation_factor));
+  start_agg_index = target_disp / (unsigned long long) (samples_in_file / agg_buffer->aggregation_factor);
+  end_agg_index = ((target_disp + target_count - 1) / (unsigned long long) (samples_in_file / agg_buffer->aggregation_factor));
   //assert(start_agg_index >= 0 && end_agg_index >= 0 && end_agg_index >= start_agg_index);
 
   /*
@@ -991,7 +991,7 @@ PIDX_return_code PIDX_global_agg_buf_create(PIDX_global_agg_id agg_id, Agg_buffe
             int l = pow(2, ((int)log2(agg_buffer->file_number * agg_id->idx->blocks_per_file)));
             adjusted_file_index = (l * (agg_id->idx_d->idx_count[0] * agg_id->idx_d->idx_count[1] * agg_id->idx_d->idx_count[2]) + ((agg_buffer->file_number * agg_id->idx->blocks_per_file) - l) + (agg_id->idx_d->color * l)) / agg_id->idx->blocks_per_file;
 
-            uint64_t sample_count = global_block_layout->block_count_per_file[agg_buffer->file_number] * agg_id->idx_d->samples_per_block / agg_buffer->aggregation_factor;
+            unsigned long long sample_count = global_block_layout->block_count_per_file[agg_buffer->file_number] * agg_id->idx_d->samples_per_block / agg_buffer->aggregation_factor;
 
             int total_chunk_size = agg_id->idx->chunk_size[0] * agg_id->idx->chunk_size[1] * agg_id->idx->chunk_size[2] * agg_id->idx->chunk_size[3] * agg_id->idx->chunk_size[4];
 
@@ -1111,9 +1111,9 @@ static PIDX_return_code create_file_zero_buffer(PIDX_global_agg_id agg_id, PIDX_
   MPI_Comm_rank(agg_id->comm, &rank);
 #endif
 
-  int64_t send_index = 0;
+  unsigned long long send_index = 0;
   int p = 0, i = 0, bytes_for_datatype = 0, v = 0;
-  //int64_t samples_per_file = (int64_t) agg_id->idx_d->samples_per_block * agg_id->idx->blocks_per_file;
+  //unsigned long long samples_per_file = (unsigned long long) agg_id->idx_d->samples_per_block * agg_id->idx->blocks_per_file;
   int count = 0;
   int chunk_size = agg_id->idx->chunk_size[0] * agg_id->idx->chunk_size[1] * agg_id->idx->chunk_size[2] * agg_id->idx->chunk_size[3] * agg_id->idx->chunk_size[4];
   int buffer_count = 0;
@@ -1512,8 +1512,8 @@ PIDX_return_code PIDX_global_agg(PIDX_global_agg_id agg_id, Agg_buffer agg_buffe
   if (use_fixed_accuracy_compression == 1)
   {
   /* Scan and "pack" the aggregation buffer to get rid of the metadata */  
-  uint64_t start_block = 0;  
-  uint64_t compressed_size = 0;
+  unsigned long long start_block = 0;
+  unsigned long long compressed_size = 0;
   int sample_size = agg_id->idx->variable[agg_buffer->var_number]->bits_per_value / 8;
   uint32_t modified_sample_size = sample_size + 2*sizeof(uint32_t) + sizeof(uint8_t);
   uint32_t total_sample_counts = 0;
@@ -1536,12 +1536,12 @@ PIDX_return_code PIDX_global_agg(PIDX_global_agg_id agg_id, Agg_buffer agg_buffe
   uint32_t samples_per_block = agg_id->idx_d->samples_per_block; // TODO
   uint32_t num_idx_blocks = total_sample_counts / samples_per_block;
   agg_buffer->num_idx_blocks = num_idx_blocks;
-  agg_buffer->compressed_block_size = malloc(sizeof(uint64_t) * num_idx_blocks);    
+  agg_buffer->compressed_block_size = malloc(sizeof(unsigned long long) * num_idx_blocks);
   
   // allocate a new buffer to copy the data over
   unsigned char* compressed_buffer = malloc(compressed_size);
   
-  uint64_t bytes_copied = 0;
+  unsigned long long bytes_copied = 0;
   start_block = 0;
   uint32_t compressed_block_size = 0;
   uint32_t current_sample = 0;
